# backend/docs/swagger.yaml
definitions:
  application.SavedAddress:
    properties:
      address:
        type: string
      addressId:
        type: string
    type: object
  application.ScheduleItem:
    properties:
      building:
        type: string
      daysOfWeek:
        items:
          type: string
        type: array
      endTime:
        type: string
      name:
        type: string
      room:
        type: string
      scheduleId:
        type: string
      startTime:
        type: string
      type:
        type: string
    type: object
  application.SearchHistoryItem:
    properties:
      locations:
        type: string
      query:
        type: string
      searchId:
        type: string
      timestamp:
        type: string
    type: object
  application.User:
    properties:
      email:
        type: string
      firstName:
        type: string
      lastName:
        type: string
      password:
        type: string
      userId:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: This is the Concordia Waze API server.
  title: Concordia Waze API
  version: "1.0"
paths:
...
  /users/{userId}/profile:
    get:
      description: Get a user profile by ID
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/application.User'
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user profile
      tags:
      - users
    post:
      consumes:
      - application/json
      description: Create or overwrite a user profile
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: User profile
        in: body
        name: profile
        required: true
        schema:
          $ref: '#/definitions/application.User'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create user profile
      tags:
      - users
  /users/{userId}/savedAddresses:
    get:
      description: Get favorite addresses for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/application.SavedAddress'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get saved addresses
      tags:
      - addresses
    post:
      consumes:
      - application/json
      description: Add a favorite address for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Saved address
        in: body
        name: address
        required: true
        schema:
          $ref: '#/definitions/application.SavedAddress'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add favorite
      tags:
      - addresses
  /users/{userId}/savedAddresses/{addressId}:
    delete:
      description: Delete a saved address for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Address ID
        in: path
        name: addressId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete saved address
      tags:
      - addresses
    put:
      consumes:
      - application/json
      description: Update a saved address for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Address ID
        in: path
        name: addressId
        required: true
        type: string
      - description: Address updates
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update saved address
      tags:
      - addresses
  /users/{userId}/schedule:
    get:
      description: Get schedule items for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/application.ScheduleItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get user schedule
      tags:
      - schedule
    post:
      consumes:
      - application/json
      description: Add a schedule item for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Schedule item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/application.ScheduleItem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add schedule item
      tags:
      - schedule
  /users/{userId}/schedule/{scheduleId}:
    delete:
      description: Delete a schedule item for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Schedule ID
        in: path
        name: scheduleId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete schedule item
      tags:
      - schedule
    put:
      consumes:
      - application/json
      description: Update a schedule item for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Schedule ID
        in: path
        name: scheduleId
        required: true
        type: string
      - description: Schedule updates
        in: body
        name: updates
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update schedule item
      tags:
      - schedule
  /users/{userId}/search-history:
    delete:
      description: Delete all search history items for a user
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Clear search history
      tags:
      - search
    get:
      description: Get a user's search history
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - default: 50
        description: Limit results
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/application.SearchHistoryItem'
            type: array
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get search history
      tags:
      - search
    post:
      consumes:
      - application/json
      description: Add a search query to a user's history
      parameters:
      - description: User ID
        in: path
        name: userId
        required: true
        type: string
      - description: Search item
        in: body
        name: item
        required: true
        schema:
          $ref: '#/definitions/application.SearchHistoryItem'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Add search history
      tags:
      - search
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"